fake_blockchain_connector/blockchain_votes_processor/requirements.txt
aiohttp
aio_pika
exonum-python-client @ git+https://github.com/PeterZhizhin/exonum-python-client
pynacl
pytest
